Add programs based on actual unrolled circuit.
The existing unrolled.txt is very different from this
design. These programs are based on the unrolled circuit
diagram in the thesis, which is functionally equivalent
to the PQR system but with a different cause-effect
structure.

unrolled-circuit.txt is the most direct translation of
that circuit. It is cleanly black-boxable with bits
1, 4, and 6 forming a complex with Phi=1.

unrolled-short-circuit.txt is a "short-circuited"
version of the program, stopping after it finds a state
bit being 1. This is not cleanly black-boxable, but
interestingly has the property that it forces the
system into a well-formed state with regard to the
target PQR system.