Eliminate extra I bit and run with all states.
Programs of length 2^N were producing an extra unused
bit of I in the micro causal model, incorrectly resulting in
zero phi.

And since states are explored in little-endian order, exploring
all the states starts with the same well-formed states that the
code was previously limiting to, so the user can kill the
process when they're satisfied with the output.